Pass existing PDO object to PhpRbac.
Some users might find it userful to use an existing PDO object to
connect to the PhpRbac database.

This is useful for applications that use a single PDO object for all
database connections and PDO connections that use more advanced
connection options such as MySQL SSL support. Discussed in the manual
here: http://php.net/pdo_mysql

This commit adds the functionality to allow PhpRbac to use an existing
PDO connection to access the database.

The change is backward compattible with the current way of defining the
database in the databse.config file. To make use of an existing PDO
connection, simply pass an array as the second argument of Rbac's
constructor, with key's 'adapter' being the PDO object and 'tablePrefix'
being your PhpRbac database prefix.