Removing auto initializations of database.
As per the discussion on the OWASP/rbac GitLab
(https://github.com/OWASP/rbac/pull/40) it seems that the database
connection should not be initialized in Jf.php.

I've made the necessary changes to allow for the database connection and
table prefix to be passed to the constructor.

I've also modified Jf.php to ensure that the desired
group_concat_max_len settting is set before database pathId function is
run, buy checking the database, rather than trusting the class to have
set it. A happy side-effect of this is that the desired size can now be
easilly modified simply by changing the static member of the Jf class.